Update main.sh
In this updated version, the function checks if the first argument is "noninteractive". If it is, instead of prompting the user for input, it sources a temporary file passed as the second argument to get the values for YOURDOMAIN, YOUR_PROJECT_ID, YOUR_STATIC_IP, and YOUR_REGION.

If the script is running in interactive mode (i.e., without the "noninteractive" argument), the function behaves as before, prompting the user to populate the input.

Sorta obvious but this was added so the script can be used by other scripts to monitor VMs.